// app-ui/src/app/backend/model/datasetListItem.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DatasetListItem { 
    /**
     * The dataset id assigned by Process mining process.
     */
    datasetid?: string;
    /**
     * The name of the dataset
     */
    name?: string;
    /**
     * The csv file name.
     */
    fileName?: string;
    /**
     * The csv full path.
     */
    filePath?: string;
    /**
     * The dataset description.
     */
    description?: string;
    /**
     * The dataset created datetime.
     */
    createdDate?: number;
    /**
     * The final status of dataset preview which is returned by process mining.
     */
    status?: string;
    /**
     * The datetime of last data preview.
     */
    lastPreviewDate?: number;
    /**
     * The type of the dataset.
     */
    type?: string;
    /**
     * The error message.
     */
    message?: string;
}