// app-ui/src/app/backend/model/dataset.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Schema } from './schema';
import { PreviewStatus } from './previewStatus';
import { RedisFileInfo } from './redisFileInfo';
import { PublishedViews } from './publishedViews';
import { DatasetSource } from './datasetSource';


export interface Dataset { 
    /**
     * The dataset id.
     */
    Dataset_Id?: string;
    /**
     * The name of the dataset
     */
    Dataset_Name?: string;
    /**
     * The description of the dataset
     */
    Dataset_Description?: string;
    /**
     * The columns of the csv file
     */
    schema?: Array<Schema>;
    Dataset_Source?: DatasetSource;
    /**
     * The dataset created datetime.
     */
    createdDate?: number;
    /**
     * The dateset updated datetime.
     */
    updatedDate?: number;
    /**
     * The final data preview status
     */
    status?: string;
    /**
     * The last data preview datetime.
     */
    lastPreviewDate?: number;
    /**
     * The published view data source path.
     */
    PublishedView?: string;
    previewStatus?: PreviewStatus;
    TdvView?: PublishedViews;
    CsvFile?: RedisFileInfo;
    /**
     * The type of dataset
     */
    type?: string;
    /**
     * The method of csv file is set.
     */
    csvMethod?: string;
    /**
     * Whether the dataset is marked as deleted.
     */
    deleted?: boolean;
}